Split classes into 3 fiiles
